refactor : CalulatorDelegate 프로토콜 추가 및 응집성을 위한 코드 이동
